Split gmx library in several library: gmxsys, topsys, itp, rtp, ...
