+++ gromacs_py/gmx/gmxsys.py
@@ -10,19 +10,20 @@ import copy
 import logging
 import sys
 
-from shutil import copy as shutil_copy
-
 from os_command_py import os_command
 from pdb_manip_py import pdb_manip
 from pdb_manip_py import pdb2pqr
 
+
+from .topsys import TopSys
+from .itp import Itp
 
 # Logging
 logger = logging.getLogger(__name__)
 
 # In case gmx5 is launched as main, relative import will failed
 try:
-    from .tools import monitor, ambertools
+    from ..tools import monitor, ambertools
 except ImportError:
     logger.info("Relative import from .tools fails,"
                 " use absolute import instead")
@@ -50,6 +51,11 @@ def show_log(pdb_manip_log=True):
     logger.setLevel(logging.INFO)
     # Add sys.sdout as handler
     logger.addHandler(logging.StreamHandler(sys.stdout))
+
+    # Show log of top sys:
+    from . import topsys
+    topsys.show_log()
+
     if pdb_manip_log:
         # Show pdb_manip Logs:
         pdb_manip.show_log()
@@ -106,7 +112,7 @@ logger.info('FORCEFIELD_PATH_LIST = {}'.format(FORCEFIELD_PATH_LIST))
 
 # Test folder path
 GMX_LIB_DIR = os.path.dirname(os.path.abspath(__file__))
-TEST_PATH = os.path.join(GMX_LIB_DIR, "test_files/")
+TEST_PATH = os.path.join(GMX_LIB_DIR, "../test_files/")
 
 
 # Global variable
@@ -121,1313 +127,6 @@ HA_NAME += ['O5\'', 'C5\'', 'C4\'', 'O4\'', 'C1\'', 'N1', 'C6',
             'CG2', 'C5', 'C4', 'N4', 'N3', 'C2', 'O2', 'C3\'',
             'C2\'', 'O3\'', 'P', 'O1P', 'O2P', 'N9', 'C8', 'N7',
             'O6', 'N2', 'C7', 'N6', 'O4']
-
-# Boltzmann Constant
-KB = 8.31446261815324
-
-
-class TopSys:
-    """Topologie base on gromacs .top :
-    #include forcefield
-
-    All name and full path of itp are save
-    [ system ] -> Name
-    [ molecules ] -> Composition
-
-    :param path: topologie file path
-    :type path: str
-
-    :param forcefield: name of the focefield
-    :type forcefield: str
-
-    :param itp_list: list of the itp object
-    :type itp_list: list
-
-    :param mol_comp: molecular composition
-    :type mol_comp: list
-
-    :param name: name of the system
-    :type tpr: str
-
-    :param folder: path of the top file folder
-    :type folder: str
-
-    :param include_itp: Flag indicating if the topologie include a molecule
-        topologie
-    :type include_itp: bool
-    """
-
-    def __init__(self, top_in):
-
-        self.path = os_command.full_path_and_check(top_in)
-        self.forcefield = None
-        self.itp_list = []
-        self.mol_comp = []
-        self.name = ""
-        self.folder = os_command.get_directory(top_in)
-        self.include_itp = False
-        # Intermolecular restraints (only for Free ener):
-        self.inter_interact = False
-        self.inter_bond_list = []
-        self.inter_angl_list = []
-        self.inter_dihe_list = []
-
-        self.read_file(top_in)
-        if self.include_itp:
-            logger.info("Rewrite topologie: {}".format(top_in))
-            self.write_file(top_in)
-
-    def read_file(self, top_in):
-
-        field = None
-        ifdef = False
-        bond_list = []
-        angle_list = []
-        dihed_list = []
-
-        with open(top_in) as topfile:
-            for line in topfile:
-                # print("line: ",line)
-                # Check #ifdef field:
-                if line[:6] == '#ifdef':
-                    ifdef = True
-                if line[:6] == '#endif':
-                    ifdef = False
-                # Check include files
-                if not ifdef and line.startswith('#include'):
-                    # get the file name to include:
-                    file_name = line.split()[1][1:-1]
-                    # remove '"' , path and .itp:
-                    include = (file_name.split("/")[-1]).split(".")[0]
-
-                    # Look if the itp is in the top path:
-                    # print("Path 1: ",self.folder+"/"+file_name)
-                    # print("Path 2: ",FORCEFIELD_PATH_LIST+"/"+file_name)
-
-                    itp_found = False
-                    if os_command.check_file_exist(
-                            os.path.join(self.folder, file_name)):
-                        path = os.path.abspath(os.path.join(self.folder,
-                                                            file_name))
-                        itp_found = True
-                    else:
-                        for forcefield in FORCEFIELD_PATH_LIST:
-                            if os_command.check_file_exist(
-                                    os.path.join(forcefield, file_name)):
-                                path = os.path.abspath(
-                                    os.path.join(forcefield, file_name))
-                                itp_found = True
-                                break
-                    if not itp_found:
-                        raise IOError('Itp ' + file_name + ' not found')
-
-                    # print("name =", include, "fullname =", file_name,
-                    # "path =",path)
-                    if include == "forcefield":
-                        self.forcefield = {'name': file_name.split('.')[0],
-                                           'fullname': file_name,
-                                           'path': path}
-                    else:
-                        self.add_mol_itp(path)
-                # Check field
-                elif not ifdef and line.startswith("["):
-                    # Remove space and [ ]
-                    field = line.strip()[1:-1].strip()
-                    # print(field)
-                    # As intermolecular_inter field is empty
-                    # (filled with bonds, angles, ...)
-                    # It has to be checked now:
-                    if field == 'intermolecular_interactions':
-                        self.inter_interact = True
-                    continue
-                # Check in the field :
-                elif (not ifdef and not line[0].startswith(";") and
-                      line.strip() != ""):
-                    if field == 'moleculetype':
-                        # in the case where mol param are present in the top
-                        # file, convert the top to an itp
-                        # If mol_name variable is defined, give to the mol
-                        # param this name
-                        name_itp = os.path.basename(top_in)[:-3] + "itp"
-
-                        logger.info("Molecule topologie present in {} "
-                                    ", extract the topologie in a separate"
-                                    " file: {}".format(top_in, name_itp))
-                        # Store and write the itp file:
-                        top_itp = Itp(name=name_itp, fullname=name_itp,
-                                      path=os.path.abspath(top_in))
-                        top_itp.write_file(os.path.join(self.folder, name_itp))
-                        top_itp.display()
-                        # Add the itp to the itp_list
-                        self.add_mol_itp(os.path.join(self.folder, name_itp))
-                        self.include_itp = True
-                    # Name of the system
-                    elif field == 'system':
-                        self.name = line.strip()
-                    # Molecule composition of the system
-                    elif field == 'molecules':
-                        line_list = line.strip().split()
-                        self.mol_comp.append({'name': line_list[0],
-                                              'num': line_list[1]})
-                    # Following command concern
-                    # intermolecular interactions
-                    elif field == 'bonds':
-                        bond_list.append(line.strip().split())
-                    elif field == 'angles':
-                        angle_list.append(line.strip().split())
-                    elif field == 'dihedrals':
-                        dihed_list.append(line.strip().split())
-        if self.inter_interact:
-            self.add_intermolecular_restr(bond_list=bond_list,
-                                          angle_list=angle_list,
-                                          dihed_list=dihed_list)
-
-    def display(self):
-        if self.forcefield:
-            print("Forcefield include :\n", self.forcefield['name'])
-        else:
-            print('No forcefield defined !')
-        for itp in self.itp_list:
-            itp.display()
-        print("Mol List:")
-        for mol in self.mol_comp:
-            print("   * {} {}".format(mol['num'], mol['name']))
-
-        print("Mol Name:\n", self.name)
-
-    def write_file(self, top_out):
-        filout = open(top_out, 'w')
-
-        filout.write("; Topologie file created by " + __name__ + "\n")
-        if self.forcefield:
-            filout.write("; Forcefield: \n")
-            filout.write("#include \"" + self.forcefield['fullname'] + "\"\n")
-        # print include files
-        filout.write("\n; Itp to include: \n")
-        for itp in self.itp_list:
-            filout.write("#include \"" + itp.fullname + "\"\n")
-            # Check if the include is in the ff folder:
-            # if self.forcefield['fullname'].split("/")[0] == \
-            #       itp.fullname.split("/")[0]:
-            #    filout.write("#include \""+itp.fullname+"\"\n")
-            # else:
-            #    filout.write("#include \""+itp.fullname+"\"\n")
-        filout.write("\n[ system ]\n; Name\n")
-        filout.write(self.name + "\n")
-        filout.write("\n[ molecules ]\n; Compound        #mols\n")
-        for mol in self.mol_comp:
-            filout.write(mol['name'] + " \t" + mol['num'] + "\n")
-        filout.write("\n")
-        # INTER
-        if self.inter_interact:
-            filout.write("[ intermolecular_interactions ]\n")
-            if self.inter_bond_list:
-                filout.write("\n[ bonds ]\n;  ai    aj    type   "
-                             " bA    kA    bB    kB\n")
-                for param in self.inter_bond_list:
-                    filout.write(
-                        "{:>6}{:>6}{:>6}{:>13}{:>13}{:>13}{:>13}\n".format(
-                            param['ai'], param['aj'], param['type'],
-                            param['rA'], param['kA'],
-                            param['rB'], param['kB']))
-            if self.inter_angl_list:
-                filout.write("\n[ angles ]\n;  ai    aj    ak  "
-                             "  type    thA    kA    thB    kB\n")
-                for param in self.inter_angl_list:
-                    filout.write("{:>6}{:>6}{:>6}{:>6}{:>13}{:>13}{:>13}"
-                                 "{:>13}\n".format(
-                                    param['ai'], param['aj'],
-                                    param['ak'], param['type'],
-                                    param['thA'], param['kA'],
-                                    param['thB'], param['kB']))
-            if self.inter_dihe_list:
-                filout.write("\n[ dihedrals ]\n;  ai    aj    ak    al"
-                             "    type    phiA    kA    phiB    kB\n")
-                for param in self.inter_dihe_list:
-                    filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}{:>13}{:>13}"
-                                 "{:>13}{:>13}\n".format(
-                                    param['ai'], param['aj'],
-                                    param['ak'], param['al'],
-                                    param['type'],
-                                    param['thA'], param['kA'],
-                                    param['thB'], param['kB']))
-
-        filout.close()
-        self.copy_dependancies(os_command.get_directory(top_out))
-
-    def charge(self):
-        """Get the charge of the system
-        """
-
-        self._charge = 0
-        # Look for mol name
-        for mol in self.mol_comp:
-            name = mol['name']
-            num = int(mol['num'])
-            # print("Mol top info: ",name,num)
-            # Search mol name in itp:
-            for local_itp in self.itp_list:
-                itp_charge = local_itp.charge(name)
-                if itp_charge is not None:
-                    logger.debug("Get charge of {} : {} total charge:"
-                                 " {}".format(name, itp_charge,
-                                              itp_charge * num))
-                    self._charge += num * itp_charge
-                    break
-        return self._charge
-
-    def prot_res_num(self, selection="Protein"):
-        """Compute the residue number of a selection
-        """
-
-        self.res_num = 0
-        # Look for mol name
-        for mol in self.mol_comp:
-            name = mol['name']
-            num = int(mol['num'])
-            logger.info("{} : {}".format(name, num))
-            if name.find(selection) != -1:
-                # Search mol name in itp:
-                for itp in self.itp_list:
-                    # print(itp.name)
-                    itp_res_num = itp.res_num(name)
-                    if itp_res_num is not None:
-                        logger.info("Get Res num of {} : {}\n"
-                                    "Total number of residue: {}".format(
-                                        name, itp_res_num, num * itp_res_num))
-                        self.res_num += num * itp_res_num
-                        break
-        return self.res_num
-
-    def mol_num(self, name):
-        """Get the number of the molecule "name"
-        """
-
-        mol_num = 0
-
-        for mol in self.mol_comp:
-            if mol['name'] == name:
-                mol_num = mol_num + int(mol['num'])
-        return mol_num
-
-    def add_posre(self, posre_name="POSRE_CA",
-                  selec_dict={'atom_name': ['CA']},
-                  fc=[1000, 1000, 1000]):
-        """Add position restraint based on the selection for each itp
-        """
-        for mol in self.mol_comp:
-            for itp in self.itp_list:
-                itp.add_posre(mol_name=mol['name'], posre_name=posre_name,
-                              selec_dict=selec_dict, fc=fc)
-
-    def add_mol_itp(self, mol_itp_file):
-        """Add a molecule itp in the topologie itp_list.
-        """
-        fullname = (mol_itp_file.split("/")[-1])
-        include = fullname.split(".")[0]
-        path = os_command.full_path_and_check(mol_itp_file)
-        mol_itp = Itp(name=include, fullname=fullname, path=path)
-
-        # 1. Get new mol name:
-        mol_name_list = []
-        for top_mol in mol_itp.top_mol_list:
-            mol_name_list.append(top_mol.name)
-
-        # 2. Check if it already present
-        present = False
-        for itp in self.itp_list:
-            for top_mol in itp.top_mol_list:
-                if top_mol.name in mol_name_list:
-                    present = True
-                    break
-            if present:
-                break
-
-        # Add the itp if not present
-        if not present:
-            self.itp_list.append(mol_itp)
-
-    def add_mol(self, mol_name, mol_itp_file, mol_num):
-        """Add a molecule in the topologie (composition and itp_list)
-        """
-
-        logger.info("Add {} mol {}".format(mol_num, mol_itp_file))
-        self.mol_comp.append({'name': mol_name, 'num': str(mol_num)})
-        self.add_mol_itp(mol_itp_file)
-
-    def add_atomtypes(self, new_atomtypes):
-        """ Add atomtypes in a topologie.
-        """
-
-        # check if and atomtypes file exists:
-        # self.display()
-        atom_type = False
-        for itp_file in self.get_include_no_posre_file_list():
-            if itp_file.split("/")[-1].endswith('atomtypes.itp'):
-                # print('atomtypes file present', itp_file)
-                atom_type = True
-                atomtype_path = itp_file
-                break
-
-        # If it doesn't exist create it
-        if not atom_type:
-            fullname = (new_atomtypes.split("/")[-1])
-            include = fullname.split(".")[0]
-            path = os_command.full_path_and_check(new_atomtypes)
-            atomtype_itp = Itp(name=include, fullname=fullname, path=path)
-
-            self.itp_list = [atomtype_itp] + self.itp_list
-        # If it does exist add the new atomtypes in the first one:
-        else:
-            # First extract old atom types:
-            field = None
-            atom_dict = {}
-            name_list = []
-            with open(atomtype_path) as file:
-                for line in file:
-                    if line.strip().startswith("["):
-                        # Remove space and [ ], remove also comments
-                        field = line.replace(" ", "").split("]")[0][1:]
-                        continue
-
-                    if (line[0] != ";" and line[0] != "#" and
-                            line.strip() != ""):
-                        # Remove commentary in the line
-                        line_comment = line.split(';')
-                        line_list = line_comment[0].split()
-
-                        if field == 'atomtypes':
-                            name_list.append(line_list[0])
-                            name, bond_type, mass, charge, ptype, sigma, \
-                                epsilon = line_list[:7]
-                            atom_dict[name] = {'bond_type': bond_type,
-                                               'mass': float(mass),
-                                               'charge': float(charge),
-                                               'ptype': ptype,
-                                               'sigma': float(sigma),
-                                               'epsilon': float(epsilon)}
-                            name_list.append(name)
-
-            # Second extract new atom types and
-            # check if they are already present:
-            field = None
-            new_atom_dict = {}
-            with open(new_atomtypes) as file:
-                for line in file:
-                    if line.strip().startswith("["):
-                        # Remove space and [ ], remove also comments
-                        field = line.replace(" ", "").split("]")[0][1:]
-                        continue
-
-                    if (line[0] != ";" and line[0] != "#" and
-                            line.strip() != ""):
-                        # Remove commentary in the line
-                        line_comment = line.split(';')
-                        line_list = line_comment[0].split()
-
-                        if field == 'atomtypes':
-                            name, bond_type, mass, charge, ptype, sigma, \
-                                epsilon = line_list[:7]
-                            local_dict = {'bond_type': bond_type,
-                                          'mass': float(mass),
-                                          'charge': float(charge),
-                                          'ptype': ptype,
-                                          'sigma': float(sigma),
-                                          'epsilon': float(epsilon)}
-                            if name not in name_list:
-                                new_atom_dict[name] = local_dict
-                            else:
-                                # Check if the new values are the same:
-                                if local_dict != atom_dict[name]:
-                                    logger.warning(
-                                        'Atom types parameters for {} are '
-                                        'different in {} and {}. Only one '
-                                        'version is kept !!!'.format(
-                                            name, atomtype_path,
-                                            new_atomtypes))
-
-            # Finally append the new param in the old one
-            with open(atomtype_path, 'a') as file:
-                for name, atom_dict in new_atom_dict.items():
-                    file.write(' {:3}      {:3}         {:.5f}  {:.5f}'
-                               '   {}     {:.5e}   {:.5e}\n'.format(
-                                name,
-                                atom_dict['bond_type'],
-                                atom_dict['mass'],
-                                atom_dict['charge'],
-                                atom_dict['ptype'],
-                                atom_dict['sigma'],
-                                atom_dict['epsilon']))
-
-    def add_intermolecular_restr(self, bond_list=[],
-                                 angle_list=[], dihed_list=[]):
-        """ Add inter molecular restraints in topologie file
-        """
-
-        self.inter_interact = True
-        for bond in bond_list:
-            ai, aj, funct, rA, kA, rB, kB = bond
-            self.inter_bond_list.append({'ai': ai, 'aj': aj,
-                                         'type': funct,
-                                         'rA': rA, 'kA': kA,
-                                         'rB': rB, 'kB': kB})
-        for angle in angle_list:
-            ai, aj, ak, funct, thA, kA, thB, kB = angle
-            self.inter_angl_list.append({'ai': ai, 'aj': aj,
-                                         'ak': ak, 'type': funct,
-                                         'thA': thA, 'kA': kA,
-                                         'thB': thB, 'kB': kB})
-        for dihed in dihed_list:
-            ai, aj, ak, al, funct, thA, kA, thB, kB = dihed
-            self.inter_dihe_list.append({'ai': ai, 'aj': aj,
-                                         'ak': ak, 'al': al,
-                                         'type': funct,
-                                         'thA': thA, 'kA': kA,
-                                         'thB': thB, 'kB': kB})
-
-    def change_mol_num(self, mol_name, mol_num):
-        """ Update molecule number.
-        And remove multiple molecule definition if they are consecutive.
-        """
-
-        for i, mol in enumerate(self.mol_comp):
-
-            if mol['name'] == mol_name:
-                mol['num'] = str(mol_num)
-                # Find other mol_name and remove them:
-                to_remove_index_list = []
-                for j in range(i + 1, len(self.mol_comp)):
-                    if self.mol_comp[j]['name'] == mol_name:
-                        to_remove_index_list.append(j)
-                    else:
-                        break
-                if len(to_remove_index_list) > 0:
-                    for index in sorted(to_remove_index_list, reverse=True):
-                        del self.mol_comp[index]
-                break
-        logger.info(self.mol_comp)
-
-    def get_include_file_list(self):
-        file_list = []
-        # print("\n\nIncluded files: ")
-        for itp in self.itp_list:
-            if self.forcefield:
-                if (self.forcefield['fullname'].split("/")[0] !=
-                        itp.fullname.split("/")[0]):
-                    # print(itp.path)
-                    file_list.append(itp.path)
-                    file_list = file_list + itp.get_include_file_list()
-            else:
-                file_list.append(itp.path)
-                file_list = file_list + itp.get_include_file_list()
-
-        return file_list
-
-    def get_include_no_posre_file_list(self):
-        file_list = []
-        # print("\n\nIncluded files: ")
-        for itp in self.itp_list:
-            if self.forcefield:
-                if (self.forcefield['fullname'].split("/")[0] !=
-                        itp.fullname.split("/")[0]):
-                    file_list.append(itp.path)
-            else:
-                file_list.append(itp.path)
-        return file_list
-
-    def copy_top_and_dependancies(self, dest_file):
-
-        dest_folder = os_command.get_directory(dest_file)
-        logger.info("Copy topologie file and dependancies")
-
-        if self.path != os.path.abspath(dest_file):
-            shutil_copy(self.path, os.path.abspath(dest_file))
-
-        self.copy_dependancies(dest_folder)
-
-    def copy_dependancies(self, dest_folder):
-
-        dest_folder = os.path.abspath(dest_folder)
-
-        file_to_copy = self.get_include_file_list()
-        # print(file_to_copy)
-        for file in file_to_copy:
-            if os.path.abspath(os_command.get_directory(file)) !=\
-                    os.path.abspath(dest_folder):
-                # print("Copy: "+file+":to: "+dest_folder)
-                shutil_copy(file, dest_folder)
-
-    def change_mol_name(self, old_name, new_name):
-
-        # Change name in mol_comp
-        name_change_flag = False
-        for mol in self.mol_comp:
-            if mol['name'] == old_name:
-                name_change_flag = True
-                mol['name'] = new_name
-        # Change in itp :
-        for itp in self.itp_list:
-            itp.change_mol_name(old_name, new_name)
-
-        if name_change_flag:
-            self.write_file(self.path)
-
-
-class Itp:
-    """Itp topologie in gromacs format
-    May contain several molecule definition, so
-    itp class is a collection of top_mol object
-    which are indivudial molecule topologies
-    """
-
-    def __init__(self, name, fullname, path):
-        """Read an itp file and extract all the top_mol
-        """
-
-        # name is useless to fix, maybe only put shortpath and fullpath
-        self.name = name
-        self.fullname = fullname
-        self.path = path
-        self.top_mol_list = []
-        self.posres_file = []
-        self.read_file()
-
-    def read_file(self):
-
-        field = None
-        ifdef = False
-        posre_def = ""
-        local_top = None
-        local_top_flag = False
-
-        with open(self.path) as file:
-            for line in file:
-                # print("Itp line: \"{}\" ".format(line))
-                # Check posres include:
-                if line[:6] == '#ifdef':
-                    ifdef = True
-                    line_list = line.split()
-                    posre_def = line_list[1]
-                if line[:6] == '#endif':
-                    ifdef = False
-                if ifdef and line[:8] == '#include':
-                    line_list = line.split()
-                    posre_file = line_list[1][1:-1]
-                    self.posres_file.append(
-                        {'def': posre_def, 'file': posre_file})
-                if line.strip().startswith("["):
-                    # Remove space and [ ], remove also comments
-                    field = line.replace(" ", "").split("]")[0][1:]
-                    continue
-                if line[0] != ";" and line[0] != "#" and line.strip() != "":
-                    # Remove commentary in the line
-                    line_comment = line.split(';')
-                    line_list = line_comment[0].split()
-                    # print(line_list)
-
-                    if field == 'moleculetype':
-                        # Check if a top_mol already exist, if yes append it
-                        # to the top_mol_list
-                        if local_top_flag:
-                            # print("Add mol topologie",local_top.name)
-                            self.top_mol_list.append(local_top)
-                        # Create a new top_mol
-                        local_top = TopMol(line_list[0], line_list[1])
-                        local_top_flag = True
-
-                    elif field == 'atoms':
-                        atom_num = int(line_list[0])
-                        atom_type = line_list[1]
-                        # In rare cases the residue num is not an integer
-                        # With duplicate residues like : eg "184A"
-                        try:
-                            res_num = int(line_list[2])
-                        except ValueError:
-                            res_num = line_list[2]
-                        res_name = line_list[3]
-                        atom_name = line_list[4]
-                        charge_num = int(line_list[5])
-                        charge = float(line_list[6])
-                        if len(line_list) > 7:
-                            mass = float(line_list[7])
-                        else:
-                            mass = None
-                        atom = {"num": atom_num,
-                                "atom_type": atom_type,
-                                "atom_name": atom_name,
-                                "res_num": res_num,
-                                "res_name": res_name,
-                                "charge_num": charge_num,
-                                "charge": charge,
-                                "mass": mass}
-                        local_top.atom_dict[atom_num] = atom
-                    # Read only indexes not the parameters (c0, c1 ...)
-                    # May need some modifications :
-                    elif field == 'bonds':
-                        # for i in range(3):
-                        #    print("val:{} #{}#".format(i, line[i*5:(i+1)*5]))
-                        # ai, aj, funct = [int(line[i*5:(i+1)*5]) for i in
-                        # range(3)]
-                        ai, aj, funct = [int(col) for col in line_list[:3]]
-                        if len(line_list) < 5:
-                            r = k = ''
-                        else:
-                            r, k = line_list[3:5]
-                        local_top.bond_list.append({'ai': ai, 'aj': aj,
-                                                    'funct': funct, 'r': r,
-                                                    'k': k})
-                    elif field == 'constraints':
-                        ai, aj, funct = [int(col) for col in line_list[:3]]
-                        local_top.cons_list.append({'ai': ai, 'aj': aj,
-                                                    'funct': funct})
-                    elif field == 'pairs':
-                        ai, aj, funct = [int(col) for col in line_list[:3]]
-                        local_top.pair_list.append({'ai': ai, 'aj': aj,
-                                                    'funct': funct})
-                    elif field == 'angles':
-                        ai, aj, ak, funct = [int(col) for col in line_list[:4]]
-                        if len(line_list) < 6:
-                            theta = cth = ''
-                        else:
-                            theta, cth = line_list[4:6]
-                        local_top.angl_list.append({'ai': ai, 'aj': aj,
-                                                    'ak': ak, 'funct': funct,
-                                                    'theta': theta,
-                                                    'cth': cth})
-                    elif field == 'dihedrals':
-                        ai, aj, ak, al, funct = [int(col) for col in
-                                                 line_list[:5]]
-                        if len(line_list) < 7:
-                            phase = kd = pn = ''
-                        else:
-                            phase, kd = line_list[5:7]
-                            pn = int(line_list[7])
-                        local_top.dihe_list.append({'ai': ai, 'aj': aj,
-                                                    'ak': ak, 'al': al,
-                                                    'funct': funct,
-                                                    'phase': phase, 'kd': kd,
-                                                    'pn': pn})
-                    elif field == 'virtual_sites3':
-                        ai, aj, ak, al, funct = [
-                            int(col) for col in line_list[:5]]
-                        local_top.vs3_list.append({'ai': ai, 'aj': aj,
-                                                   'ak': ak, 'al': al,
-                                                   'funct': funct})
-                    elif field == 'cmap':
-                        ai, aj, ak, al, am, funct = [
-                            int(col) for col in line_list[:6]]
-                        local_top.cmap_list.append({'ai': ai, 'aj': aj,
-                                                    'ak': ak, 'al': al,
-                                                    'am': am, 'funct': funct})
-                    elif field == 'virtual_sites4':
-                        ai, aj, ak, al, am, funct = [
-                            int(col) for col in line_list[:6]]
-                        local_top.vs4_list.append({'ai': ai, 'aj': aj,
-                                                   'ak': ak, 'al': al,
-                                                   'am': am, 'funct': funct})
-                    elif field == 'position_restraints':
-                        # With MARTINI topologie kx, ky, kz can be strings,
-                        # so no int() conversions
-                        ai, funct, kx, ky, kz = [col for col in line_list[:5]]
-                        if ifdef:
-                            local_top.if_pos_restr.append(
-                                {'def': posre_def, 'ai': int(ai),
-                                 'funct': int(funct), 'kx': kx,
-                                 'ky': ky, 'kz': kz})
-                        else:
-                            local_top.pos_restr.append(
-                                {'ai': int(ai), 'funct': int(funct), 'kx': kx,
-                                 'ky': ky, 'kz': kz})
-
-                    # else:
-                    #   raise ValueError('Unknown field : '+field)
-
-        # Needed for empty topologies like aditional ff parameters:
-        if local_top_flag:
-            self.top_mol_list.append(local_top)
-
-    def write_file(self, itp_file):
-        filout = open(itp_file, 'w')
-        filout.write("; Itp file created by " + __name__ + "\n")
-
-        for top_mol in self.top_mol_list:
-            logger.info(top_mol.name)
-            top_mol.write_file(filout)
-        for posre in self.posres_file:
-            filout.write("\n#ifdef " + posre['def'] + "\n")
-            filout.write("#include \"" + posre['file'] + "\"\n")
-            filout.write("#endif\n\n")
-
-        filout.close()
-
-    def display(self):
-        logger.info('- ITP file: {}'.format(self.name))
-        logger.info("- molecules defined in the itp file:")
-        for top_mol in self.top_mol_list:
-            logger.info("* {}".format(top_mol.name))
-
-    def charge(self, mol_name):
-        for top_mol in self.top_mol_list:
-            # print(mol_name, top_mol.name)
-            if top_mol.name == mol_name:
-                return top_mol.get_charge()
-        return None
-
-    def res_num(self, mol_name):
-        for top_mol in self.top_mol_list:
-            # print(mol_name, top_mol.name)
-            if top_mol.name == mol_name:
-                return top_mol.get_res_num()
-        return None
-
-    def add_posre(self, mol_name, posre_name, selec_dict, fc, replace=True):
-        for top_mol in self.top_mol_list:
-            # print(mol_name, top_mol.name)
-            if top_mol.name == mol_name:
-                index_posre = top_mol.get_selection_index(
-                    selec_dict=selec_dict)
-                if index_posre:
-                    # Create the posre itp file :
-                    logger.debug("Posre for : {}".format(top_mol.name))
-                    posre_file_name = os.path.abspath(
-                        os.path.join(
-                            os_command.get_directory(self.path),
-                            self.name + "_posre_" + posre_name + ".itp"))
-                    # posre_file_name = self.name+"_posre_"+posre_name+".itp"
-                    write_index_posre_file(atom_index_list=index_posre,
-                                           posre_file=posre_file_name,
-                                           type_val=1, fc=fc)
-                    # Add the posre include in the mol itp file:
-                    # Need to solve the problem of #ifdef location with .top
-                    # files containing self top
-                    if replace:
-                        # Remove previous def:
-                        new_content = ""
-                        with open(self.path) as file:
-                            posre_def = None
-                            for line in file:
-                                # print("Itp line: \"{}\" ".format(line))
-                                # Check posres include:
-                                if line[:6] == '#ifdef':
-                                    line_list = line.split()
-                                    posre_def = line_list[1]
-
-                                if posre_def != posre_name:
-                                    new_content += line
-
-                                if line[:6] == '#endif':
-                                    posre_def = None
-                        # Add the new one:
-                        new_content += '#ifdef ' + posre_name + '\n'
-                        new_content += '#include \"' +\
-                            os.path.basename(posre_file_name) + "\" \n"
-                        new_content += '#endif \n\n'
-                        # Save the file:
-                        file = open(self.path, "w")
-                        file.write(new_content)
-                        file.close()
-
-                    else:
-                        with open(self.path, 'a') as file:
-                            file.write('#ifdef ' + posre_name + '\n')
-                            # file.write('#include \"'+self.name+"_posre_"+\
-                            # posre_name+".itp\" \n")
-                            file.write(
-                                '#include \"' +
-                                os.path.basename(posre_file_name) + "\" \n")
-                            file.write('#endif \n\n')
-
-    def set_top_mol_name(self, new_name):
-        if len(self.top_mol_list) == 1:
-            self.top_mol_list[0].name = new_name
-        else:
-            raise ValueError('Cannot set top mol name with multiple top mol '
-                             'in itp')
-
-    def get_include_file_list(self):
-        file_list = []
-        for posre in self.posres_file:
-            # print(posre)
-            file_list.append(os.path.abspath(os.path.join(
-                os_command.get_directory(self.path), posre['file'])))
-        return file_list
-
-    def change_mol_name(self, old_name, new_name):
-
-        name_change_flag = False
-        for top_mol in self.top_mol_list:
-            if top_mol.name == old_name:
-                name_change_flag = True
-                top_mol.name = new_name
-        if name_change_flag:
-            self.write_file(self.path)
-
-
-class TopMol:
-    """Individual molecule topologie
-    """
-
-    def __init__(self, name, nrexcl):
-        """Read an itp file and extract [atoms] field
-        """
-        self.name = name
-        self.nrexcl = nrexcl
-        self.atom_dict = {}
-        self.bond_list = []
-        self.cons_list = []
-        self.pair_list = []
-        self.angl_list = []
-        self.dihe_list = []
-        self.cmap_list = []
-        self.vs3_list = []
-        self.vs4_list = []
-        self.pos_restr = []
-        self.if_pos_restr = []
-
-    def get_charge(self):
-        self.charge = 0
-        for atom in self.atom_dict.values():
-            self.charge += atom['charge']
-        return round(self.charge, 4)
-
-    def get_res_num(self):
-        self.res_num = -1
-        for atom in self.atom_dict.values():
-            if atom["res_num"] != self.res_num:
-                self.res_num += 1
-        return self.res_num
-
-    def get_selection_index(self, selec_dict={'atom_name': ['CA']}):
-        """Return the atom index to add posre
-        """
-
-        index_list = []
-        for index, atom in self.atom_dict.items():
-            selected = True
-            for field, selection in selec_dict.items():
-                if atom[field] not in selection:
-                    selected = False
-                    break
-            if selected:
-                index_list.append(index)
-
-        return index_list
-
-    def write_file(self, filout):
-        filout.write("\n[ moleculetype ]\n; Name            nrexcl\n")
-        filout.write("{}\t\t{}\n".format(self.name, self.nrexcl))
-        # Print Atoms field
-        filout.write("\n[ atoms ]\n;   nr       type  resnr residue  atom   "
-                     "cgnr     charge       mass  typeB    chargeB      "
-                     "massB\n")
-        tot_charge = 0
-        for atom in self.atom_dict.values():
-            tot_charge += atom['charge']
-            # print('ATOM:',atom)
-            filout.write("{:>6}{:>11}{:>7}{:>7}{:>7}{:>7}{:>11.5f}{:>11} "
-                         "  ; qtot {:<6.2f} \n".
-                         format(atom['num'],
-                                atom['atom_type'],
-                                atom['res_num'],
-                                atom['res_name'],
-                                atom['atom_name'], atom['charge_num'],
-                                atom['charge'],
-                                atom['mass'],
-                                tot_charge))
-        # Print bonds field
-        filout.write("\n[ bonds ]\n;  ai    aj funct            c0           "
-                     " c1            c2            c3\n")
-        for param in self.bond_list:
-            filout.write("{:>6}{:>6}{:>6}{:>13}{:>13}\n".format(
-                param['ai'], param['aj'], param['funct'],
-                param['r'], param['k']))
-        # Print constraints field
-        filout.write("\n[ constraints ]\n;  ai    aj funct            c0    "
-                     "        c1\n")
-        for param in self.cons_list:
-            filout.write("{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['aj'], param['funct']))
-        # Print pairs field
-        filout.write("\n[ pairs ]\n;  ai    aj funct            c0          "
-                     "  c1            c2            c3\n")
-        for param in self.pair_list:
-            filout.write("{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['aj'], param['funct']))
-        # Print angles field
-        filout.write("\n[ angles ]\n;  ai    aj    ak funct            c0   "
-                     "         c1            c2            c3\n")
-        for param in self.angl_list:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>13}{:>13}\n".format(
-                param['ai'], param['aj'],
-                param['ak'], param['funct'],
-                param['theta'], param['cth']))
-        # Print dihedrals field
-        filout.write("\n[ dihedrals ]\n;  ai    aj    ak    al funct        "
-                     "    c0            c1            c2            c3      "
-                     "      c4            c5\n")
-        for param in self.dihe_list:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}{:>13}{:>13}"
-                         "{:>6}\n".format(param['ai'], param['aj'],
-                                          param['ak'], param['al'],
-                                          param['funct'], param['phase'],
-                                          param['kd'], param['pn']))
-        # Print virtual_sites3 field
-        filout.write("\n[ cmap ]\n;  ai    aj    ak    al    am funct\n")
-        for param in self.cmap_list:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['aj'],
-                param['ak'], param['al'],
-                param['am'], param['funct']))
-        # Print virtual_sites3 field
-        filout.write("\n[ virtual_sites3 ]\n;  ai    aj    ak    al funct    "
-                     "        c0            c1\n")
-        for param in self.vs3_list:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['aj'],
-                param['ak'], param['al'],
-                param['funct']))
-        # Print virtual_sites3 field
-        filout.write("\n[ virtual_sites4 ]\n;  ai    aj    ak    al    am"
-                     " funct            c0            c1            c2\n")
-        for param in self.vs4_list:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['aj'],
-                param['ak'], param['al'],
-                param['am'], param['funct']))
-
-        # Print position restraints
-        filout.write("\n[ position_restraints ]\n;  ai    funct         kx "
-                     "         ky          kz\n")
-        for param in self.pos_restr:
-            filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}\n".format(
-                param['ai'], param['funct'],
-                param['kx'], param['ky'],
-                param['kz']))
-        if len(self.if_pos_restr) > 0:
-            def_str = ''
-            def_end_flag = False
-            for param in self.if_pos_restr:
-
-                if param['def'] != def_str:
-                    def_str = param['def']
-                    if def_end_flag:
-                        filout.write("#endif\n\n")
-
-                    def_end_flag = True
-                    filout.write("\n#ifdef " + param['def'] + "\n")
-                    filout.write("\n[ position_restraints ]\n;  ai    funct "
-                                 "        kx          ky          kz\n")
-
-                filout.write("{:>6}{:>6}{:>6}{:>6}{:>6}\n".format(
-                    param['ai'], param['funct'],
-                    param['kx'], param['ky'],
-                    param['kz']))
-            filout.write("#endif\n\n")
-
-    def delete_atom(self, index_list):
-        # Remove atom:
-        for i in index_list:
-            del self.atom_dict[i]
-
-        # Create the dict to have all atom num consecutive staring from 1
-        dict_atom_index = {}
-        for i, atom in sorted(enumerate(self.atom_dict.items())):
-            # print(i, atom)
-            dict_atom_index[atom[0]] = i + 1
-
-        # Modify all atom to have correct index
-        new_atom_dict = {}
-        for i, atom in self.atom_dict.items():
-            local_atom = atom
-            new_index = dict_atom_index[local_atom['num']]
-            local_atom['num'] = new_index
-            local_atom['charge_num'] = new_index
-            new_atom_dict[new_index] = local_atom
-
-        self.atom_dict = new_atom_dict
-
-        #  Modify all bond, cons, pair ... to have correct index:
-        new_bond_list = []
-        for i, param in enumerate(self.bond_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']]})
-                new_bond_list.append(param)
-        self.bond_list = new_bond_list
-
-        new_cons_list = []
-        for i, param in enumerate(self.cons_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']]})
-                new_cons_list.append(param)
-        self.cons_list = new_cons_list
-
-        new_pair_list = []
-        for i, param in enumerate(self.pair_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']]})
-                new_pair_list.append(param)
-        self.pair_list = new_pair_list
-
-        new_angl_list = []
-        for i, param in enumerate(self.angl_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list) or
-                    (param['ak'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']],
-                              'ak': dict_atom_index[param['ak']]})
-                new_angl_list.append(param)
-
-        self.angl_list = new_angl_list
-
-        new_dihe_list = []
-        for i, param in enumerate(self.dihe_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list) or
-                    (param['ak'] in index_list) or
-                    (param['al'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']],
-                              'ak': dict_atom_index[param['ak']],
-                              'al': dict_atom_index[param['al']]})
-                new_dihe_list.append(param)
-
-        self.dihe_list = new_dihe_list
-
-        new_vs3_list = []
-        for i, param in enumerate(self.vs3_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list) or
-                    (param['ak'] in index_list) or
-                    (param['al'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']],
-                              'ak': dict_atom_index[param['ak']],
-                              'al': dict_atom_index[param['al']]})
-                new_vs3_list.append(param)
-        self.vs3_list = new_vs3_list
-
-        new_cmap_list = []
-        for i, param in enumerate(self.cmap_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list) or
-                    (param['ak'] in index_list) or
-                    (param['al'] in index_list) or
-                    (param['am'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']],
-                              'ak': dict_atom_index[param['ak']],
-                              'al': dict_atom_index[param['al']],
-                              'am': dict_atom_index[param['am']]})
-                new_cmap_list.append(param)
-        self.cmap_list = new_cmap_list
-
-        new_vs4_list = []
-        for i, param in enumerate(self.vs4_list):
-            if not ((param['ai'] in index_list) or
-                    (param['aj'] in index_list) or
-                    (param['ak'] in index_list) or
-                    (param['al'] in index_list) or
-                    (param['am'] in index_list)):
-                param.update({'ai': dict_atom_index[param['ai']],
-                              'aj': dict_atom_index[param['aj']],
-                              'ak': dict_atom_index[param['ak']],
-                              'al': dict_atom_index[param['al']],
-                              'am': dict_atom_index[param['am']]})
-                new_vs4_list.append(param)
-        self.vs4_list = new_vs4_list
-
-    def correct_charge_type(self, forcefield, index_list=None):
-        """ Correct the charge and atom type of an itp object,
-        base on a ff `.rtp` file.
-        This is specially usefull, to correct charge of termini resiudes
-        of a cyclicized peptide.
-
-        if index_list is None, will correct all atom charges, if not, only
-        atoms listed in index_list.
-        """
-
-        # First extract charge and type from the ff .rtp file
-        forcefield_path = forcefield['path']
-        rtp_path = os.path.abspath(os.path.join(
-            os_command.get_directory(forcefield_path), 'aminoacids.rtp'))
-        if not os_command.check_file_exist(rtp_path):
-            rtp_path = os.path.abspath(os.path.join(
-                os_command.get_directory(forcefield_path), 'merged.rtp'))
-
-        logger.info('Read rtp file : {}'.format(rtp_path))
-        ff_rtp = Rtp(rtp_path)
-
-        # Correct charges and type:
-        # for atom_num, atom in self.atom_dict.items():
-        if index_list is None:
-            index_list = self.atom_dict.keys()
-
-        for atom_num in index_list:
-
-            atom = self.atom_dict[atom_num]
-
-            res_name = atom['res_name']
-            resid = atom['res_num']
-            atom_name = atom['atom_name']
-
-            # With Amber disulfide cys are called CYX in the rtp and CYS in
-            # the itp.
-            # Don't get sense but need to check the protonation state of cys
-            # By counting atoms
-            # With charmm disuldie cys is CYS2
-            if res_name == 'CYS':
-                if len(self.get_selection_index(
-                        selec_dict={'res_num': [resid],
-                                    'res_name': ['CYS']})) == 10:
-                    if forcefield['name'].startswith('amber'):
-                        res_name = 'CYX'
-                    elif forcefield['name'].startswith('charmm'):
-                        res_name = 'CYS2'
-            # print(res_name)
-            # With gromacs all histidine are called HIS !
-            if res_name == 'HIS':
-                if len(self.get_selection_index(
-                        selec_dict={'res_num': [resid],
-                                    'res_name': ['HIS']})) == 18:
-                    if forcefield['name'].startswith('amber'):
-                        res_name = 'HIP'
-                    elif forcefield['name'].startswith('charmm'):
-                        res_name = 'HSP'
-                elif len(self.get_selection_index(
-                        selec_dict={'res_num': [resid],
-                                    'atom_name': ['HE2']})) == 1:
-                    if forcefield['name'].startswith('amber'):
-                        res_name = 'HIE'
-                    elif forcefield['name'].startswith('charmm'):
-                        res_name = 'HSE'
-                else:
-                    if forcefield['name'].startswith('amber'):
-                        res_name = 'HID'
-                    elif forcefield['name'].startswith('charmm'):
-                        res_name = 'HSD'
-
-            # print(atom)
-            # print(ff_rtp.res_dict[res_name]['atom'][atom_name])
-
-            if atom_name in ff_rtp.res_dict[res_name]['atom']:
-                atom_type = \
-                    ff_rtp.res_dict[res_name]['atom'][atom_name]['type']
-                atom_charge = \
-                    ff_rtp.res_dict[res_name]['atom'][atom_name]['charge']
-                # print('Correct residue {:4} atom {:4} atom type {:4} '
-                #       'to {:4}'.format(res_name, atom['atom_name'],
-                #                        atom['atom_type'], atom_type))
-                # print('Correct charge {:4} '
-                #       'to {:4}'.format(self.atom_dict[atom_num]['charge'],
-                #                        atom_charge))
-                if atom_type != atom['atom_type']:
-                    logger.warning('Correct residue {:4} atom {:4} atom '
-                                   'type {:4} to {:4}'.format(
-                                        res_name, atom['atom_name'],
-                                        atom['atom_type'], atom_type))
-            else:
-                logger.warning('Can\'t find residue {:4} atom {:4} atom '
-                               'type {:4} parameters in forcefield'.format(
-                                    res_name, atom['atom_name'],
-                                    atom['atom_type']))
-            self.atom_dict[atom_num]['atom_type'] = atom_type
-            self.atom_dict[atom_num]['charge'] = atom_charge
-
-
-class Rtp:
-    """Individual molecule topologie
-    """
-
-    def __init__(self, path):
-        """Read an itp file and extract [atoms] field
-        """
-        self.path = path
-        self.res_dict = {}
-        self.read_file()
-
-        # Then assigne the correct atom charge and type
-
-    def read_file(self):
-
-        field = None
-        atom_dict = {}
-        bond_list = []
-        impr_list = []
-        res_name = None
-
-        with open(self.path) as file:
-            for line in file:
-                line_strip = line.strip()
-                if len(line_strip) > 0:
-                    if line_strip[0] == "[":
-                        # Remove space and [ ], remove also comments
-                        field = line.replace(" ", "").split("]")[0][1:]
-
-                        if field not in ['bondedtype', 'atoms', 'bonds',
-                                         'impropers', '']:
-
-                            if len(atom_dict) > 0:
-                                residue = {}
-                                residue['atom'] = atom_dict
-                                residue['bond'] = bond_list
-                                residue['impr'] = impr_list
-                                self.res_dict[res_name] = residue
-
-                            res_name = field
-                            atom_dict = {}
-                            bond_list = []
-                            impr_list = []
-                        continue
-
-                    if line_strip[0] != ";" and line_strip[0] != "#":
-                        line_list = line_strip.split()
-
-                        if field == 'atoms':
-                            atom_name = line_list[0]
-                            atom_type = line_list[1]
-                            atom_charge = float(line_list[2])
-                            charge_group = int(line_list[3])
-                            atom_dict[atom_name] = {
-                                "type": atom_type,
-                                "charge": atom_charge,
-                                "charge_group": charge_group}
-
-                        elif field == 'bonds':
-                            bond_list.append({'ai': line_list[0],
-                                              'aj': line_list[1]})
-
-                        elif field == 'impropers':
-                            impr_list.append({'ai': line_list[0],
-                                              'aj': line_list[1],
-                                              'ak': line_list[2],
-                                              'al': line_list[3]})
-
-# Position restraints files:
-
-
-def write_index_posre_file(atom_index_list, posre_file, type_val=1,
-                           fc=[1000, 1000, 1000]):
-    """Write a pos restraint file based on atom index list
-    """
-
-    filout = open(posre_file, 'w')
-    filout.write("; Position restraint file created by " + __name__ + "\n")
-    filout.write("[ position_restraints ]\n")
-    filout.write("; atom  type      fx      fy      fz\n")
-
-    for index in atom_index_list:
-        filout.write("{:6d}{:6d}{:6d}{:6d}{:6d} \n".format(
-            index, type_val, fc[0], fc[1], fc[2]))
-
-    filout.write("\n")
-    filout.close()
 
 
 ################################
