+++ gromacs_py/gmx/template/charmm36-jul2017.ff/forcefield.itp
@@ -1,0 +1,26 @@
+; This force field generated by charmm2gmx.py from
+; multiple charmm parameter files 
+; and multiple charmm topology files
+#define _FF_CHARMM
+
+; This GROMACS version was created from:
+; CHARMM parameters: multiple files
+; CHARMM topology: multiple files
+; Please see forcefield.doc for files and references
+[ defaults ]
+; nbfunc	comb-rule	gen-pairs	fudgeLJ	fudgeQQ
+1	2	yes	1.0	1.0
+
+#include "ffnonbonded.itp"
+#include "ffbonded.itp"
+#include "gb.itp"
+
+; Allow the user to specify CHARMM36 parameters, instead
+; Default behavior is to use CHARMM36m
+#ifdef USE_OLD_C36
+#include "old_c36_cmap.itp"
+#else
+#include "cmap.itp"
+#endif
+
+#include "nbfix.itp"
