+++ gromacs_py/gmx/template/charmm36-jul2017.ff/gb.itp
@@ -1,0 +1,55 @@
+
+[ implicit_genborn_params ]
+
+; Atom type sar		st		pi		gbr		 hct
+ NH1			0.155	1		1.028	0.17063  0.79 ; N
+ N		0.155	1	1	0.155	 0.79 ; Proline backbone N
+ NP		0.155	1	1	0.155	 0.79 ; Proline backbone N for Proline N-terminus
+ H				0.1		1		1		0.115	 0.85 ; H
+ CT1			0.180	1		1.276	0.190	 0.72 ; C
+ HB1			0.1		1		1		0.125	 0.85 ; H
+ HB2			0.1		1		1		0.125	 0.85 ; Add by Samuel
+ CT3			0.200	1		0.880	0.190	 0.72 ; C
+ HA				0.1		1		1		0.125	 0.85 ; H
+ C				0.172	1		1.554	0.1875	 0.72 ; C
+ O				0.150	1		0.926	0.148	 0.85 ; O
+ CT2			0.190	1		1.045	0.190	 0.72 ; C
+ CT2A			0.190	1		1.045	0.190	 0.72 ;  Add by Samuel
+ NC2			0.160	1		1.215	0.17063  0.79 ; N
+ HC				0.1		1		1		0.115	 0.85 ; H
+ CC				0.172	1		1.554	0.1875	 0.72 ; C
+ CD		0.172	1	1.554	0.1875	 0.72 ; carbonyl C (in COOH termini)
+ NH2			0.160	1		1.215	0.17063  0.79 ; N
+ OC				0.170	1		0.922	0.148	 0.85 ; O
+ OB		0.170	1	0.922	0.148	 0.85 ; carbonyl O (in COOH termini)
+ S				0.180	1		1.121	0.1775	 0.96 ; S
+ SM				0.180	1		1.121	0.1775	 0.96 ; disulfide sulfur
+ HS				0.1		1		1		0.125	 0.85 ; H
+ NR1			0.155	1		1.028	0.17063  0.79 ; N
+ CPH1			0.172	1		1.554	0.1875	 0.72 ; C
+ CPH2			0.180	1		1.073	0.1875	 0.72 ; C
+ HA1			0.1		1		1		0.125	 0.85 ; Add by Samuel
+ HA2			0.1		1		1		0.125	 0.85 ; Add by Samuel
+ HA3			0.1		1		1		0.125	 0.85 ; Add by Samuel
+ HR1			0.1		1		1		0.125	 0.85 ; H
+ HR2			0.1		1		1		0.125	 0.85 ; H
+ NR2			0.155	1		1.215	0.17063  0.79 ; N
+ HR3			0.1		1		1		0.115	 0.85 ; H
+ NR3			0.155	1		1.215	0.17063  0.79 ; N
+ NH3			0.160	1		1.215	0.1625	 0.79 ; N
+ HP				0.1		1		1		0.125	 0.85 ; H
+ CP3			0.190	1		1.045	0.190	 0.72 ; C
+ CP1			0.180	1		1.276	0.190	 0.72 ; C
+ CP2			0.190	1		1.045	0.190	 0.72 ; C
+ OH1			0.152	1		1.080	0.1535	 0.85 ; O
+ CY				0.172	0.012	1.554	0.1875	 0.72 ; C
+ CA				0.180	1		1.073	0.1875	 0.72 ; C
+ CAI			0.180	1		1.073	0.1875	 0.72 ; Add by Samuel
+ HP				0.1		1		1		0.125	 0.85 ; H
+ NY				0.155	1		1.028	0.17063  0.79 ; N
+ CPT			0.172	0.012	1.554	0.1875	 0.72 ; C	  
+ MNH3		0	0	0	0	 0	  ; vsite (rigid tetrahedrical NH3 group)
+ MNH2		0	0	0	0	 0	  ; vsite
+ MCH3		0	0	0	0	 0	  ; vsite (rigid CH3 group)
+ MCH3S		0	0	0	0	 0	  ; vsite
+ MW			0	0	0	0	 0	  ; vsite
