+++ gromacs_py/gmx/template/charmm36-jul2017.ff/ions.itp
@@ -1,0 +1,84 @@
+
+
+[ moleculetype ] ; added by Bjelkmar Jan 2010, from c32b1/toppar/stream/toppar_water_ions.str
+; molname	nrexcl
+OH		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge
+1		OC			1		OH			O1		 1		-1.32
+2		H			1		OH			H1		 1		 0.32
+
+[ bonds ]
+;i	j	funct	length	force.c.
+1	2	1	0.09700 456056	 ; hydroxyl bond
+
+[ moleculetype ]
+; molname	nrexcl
+LI		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		LIT			1		LI			LI		 1		1  
+
+[ moleculetype ]
+; molname	nrexcl
+NA		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		SOD			1		NA			NA		 1		1		 
+
+[ moleculetype ]
+; molname	nrexcl
+K		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		POT			1		K			K		 1		1		 
+
+[ moleculetype ]
+; molname	nrexcl
+CS		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		CES			1		Ces			Ces		 1		1		 
+
+[ moleculetype ]
+; molname	nrexcl
+CL		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		CLA			1		CL			CL		 1		-1	 
+
+[ moleculetype ]
+; molname	nrexcl
+CA		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		CAL			1		Cal			Cal		 1		2	
+
+[ moleculetype ]
+; molname	nrexcl
+MG		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge
+#ifdef MG_ALLNER
+; Use parameters from Allner et al. (DOI: 10.1021/ct3000734)
+1		MGA			1		MG			MG		 1		2
+#else
+; Standard CHARMM parameters
+1		MG			1		MG			MG		 1		2
+#endif
+
+[ moleculetype ]
+; molname	nrexcl
+ZN		1
+
+[ atoms ]
+; id	at type		res nr	residu name at name  cg nr	charge	 
+1		ZN			1		ZN			ZN		 1		2
