+++ gromacs_py/gmx/template/charmm36-jul2017.ff/merged.arn
@@ -1,0 +1,6 @@
+
+
+; atom renaming specification
+; residue	gromacs  forcefield
+   *		   H		HN
+
