+++ gromacs_py/gmx/template/charmm36-jul2017.ff/merged.c.tdb
@@ -1,0 +1,59 @@
+
+[ None ]
+
+; CHARMM CTER
+[ COO- ]
+[ replace ]
+ C		C	CC	12.011	 0.34
+ O		OT1 OC	15.9994 -0.67
+ OXT	OT2 OC	15.9994 -0.67	
+[ add ]
+ 2	8	OT	C	CA	N
+	OC	15.9994 -0.67	-1
+[ impropers ]
+ C	CA	OT2 OT1		 
+
+; CHARMM CNEU
+[ COOH ]
+[ replace ]
+ C		C	CD	12.011	 0.72
+ O		OT1 OB	15.9994 -0.55
+ OXT	OT2 OH1 15.9994 -0.61
+[ add ]
+ 2	8	OT	C	CA	N
+	OB	15.9994 -0.55	-1
+ 1	2	HT2 OT2 C	CA
+	H	1.008	0.44	-1
+[ replace ]
+ OT2	OT2 OH1		15.9994 -0.61
+[ impropers ]
+ C	CA	OT2 OT1
+
+; Created by Michel Cuendet.
+[ CT2 ]
+[ replace ]
+ C		CC		12.011	0.55
+ O		O		15.9994 -0.55
+[ add ]
+ 1	2	NT	C	CA	N
+	NH2 14.0027 -0.62	-1
+ 2	3	HT	NT	C	CA
+	H	1.008	0.00   -1
+[ replace ]
+ HT1		H		1.008	0.30	; this one is trans to O
+ HT2		H		1.008	0.32	; this one is cis to O
+[ impropers ]
+ C	NT	CA	O
+ C	CA	NT	O
+ NT C	HT1 HT2
+ NT C	HT2 HT1
+
+
+; added by Justin Lemkul 
+[ 3TER ]
+[ replace ]
+ O3'   ON5		15.9994  -0.66
+ C3'   CN7		12.011	  0.14
+[ add ]
+ 1	2  H3T	O3'  C3'  C4'
+	HN5  1.008	 0.43	-1
