+++ gromacs_py/gmx/template/charmm36-jul2017.ff/merged.hdb
@@ -1,0 +1,399 @@
+ACE	 1		 
+3	4	HH3 CH3 C	O
+ADE 10
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N9	C2'
+1	1	H8	 C8   N9   N7
+2	3	H6	 N6   C6   C5
+1	1	H2	 C2   N1   N3
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'' C2'  C1'  C3'
+1	2	H2'  O2'  C2'  C1' 
+AIB		3		
+1	1	HN	N	-C	CA
+3	4	HB1 CB1 CA	N
+3	4	HB2 CB2 CA	N
+ALA		3		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+3	4	HB	CB	CA	N
+ARG		8		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	6	HD	CD	NE	CG
+1	1	HE	NE	CD	CZ
+2	3	HH1 NH1 CZ	NE
+2	3	HH2 NH2 CZ	NE
+ARGN	8		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	6	HD	CD	NE	CG
+1	1	HE	NE	CD	CZ
+1	2	HH1 NH1 CZ	NE
+2	3	HH2 NH2 CZ	NE
+ASN		4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	3	HD2 ND2 CG	CB
+ASP		3		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+ASPP	4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	2	HD2 OD2 CG	CB
+ATP 11
+1   6   H5'  C5'  O5'  C4'
+1   6   H5'' C5'  C4'  O5'
+1   5   H4'  C4'  C5'  O4'  C3'
+1   5   H1'  C1'  O4'  N9   C2'
+1   1   H8   C8   N9   N7
+2   3   H6   N6   C6   C5
+1   1   H2   C2   N1   N3
+1   5   H3'  C3'  C4'  C2'  O3'
+1   6   H2'' C2'  C1'  C3'
+1   2   H2'  O2'  C2'  C1'
+1   2   H3T  O3'  C3'  C2'
+ADP 11
+1   6   H5'  C5'  O5'  C4'
+1   6   H5'' C5'  C4'  O5'
+1   5   H4'  C4'  C5'  O4'  C3'
+1   5   H1'  C1'  O4'  N9   C2'
+1   1   H8   C8   N9   N7
+2   3   H6   N6   C6   C5
+1   1   H2   C2   N1   N3
+1   5   H3'  C3'  C4'  C2'  O3'
+1   6   H2'' C2'  C1'  C3'
+1   2   H2'  O2'  C2'  C1'
+1   2   H3T  O3'  C3'  C2'
+CYN		3		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	SG	CA
+CYS2	3		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	SG	CA
+CYS		4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	SG	CA
+1	2	HG1 SG	CB	CA
+CYT 10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N1	C2'
+1	1	H6	 C6   N1   C5
+1	1	H5	 C5   C6   C4
+2	3	H4	 N4   C4   C5
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'' C2'  C1'  C3'
+1	1	H2'  O2'  C2'  C1'
+DA	10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N9	C2'
+1	1	H8	 C8   N9   N7
+2	3	H6	 N6   C6   C5
+1	1	H2	 C2   N1   N3
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'  C2'  C1'  C3'
+1	6	H2'' C2'  C3'  C1'
+DT	10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N1	C2'
+1	1	H6	 C6   N1   C5
+3	4	H5	 C5M  C5   C6
+1	1	H3	 N3   C4   C2
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'  C2'  C1'  C3'
+1	6	H2'' C2'  C3'  C1'
+DG	10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N9	C2'
+1	1	H8	 C8   N9   N7
+1	1	H1	 N1   C6   C2
+2	3	H2	 N2   C2   N1
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'  C2'  C1'  C3'
+1	6	H2'' C2'  C3'  C1'
+DC	10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N1	C2'
+1	1	H6	 C6   N1   C5
+1	1	H5	 C5   C6   C4
+2	3	H4	 N4   C4   C5
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'  C2'  C1'  C3'
+1	6	H2'' C2'  C3'  C1'
+D5MC 10
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N1	C2'
+1	1	H6	 C6   N1   C5
+3	4	H5M  C5M  C5   C6
+2	3	H4	 N4   C4   C5
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'  C2'  C1'  C3'
+1	6	H2'' C2'  C3'  C1'
+GLN		5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	3	HE2 NE2 CD	CG
+GLU		4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+GLUP	5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+1	2	HE2 OE2 CD	CG
+GLY		2		
+1	1	HN	N	-C	CA
+2	6	HA	CA	C	N
+GUA 10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N9	C2'
+1	1	H8	 C8   N9   N7
+1	1	H1	 N1   C6   C2
+2	3	H2	 N2   C2   N1
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'' C2'  C1'  C3'
+1	2	H2'  O2'  C2'  C1'
+HIS1	6		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD2 CD2 CG	NE2
+1	1	HE1 CE1 ND1 NE2
+1	1	HD1 ND1 CG	CE1
+HSD		6		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD2 CD2 CG	NE2
+1	1	HE1 CE1 ND1 NE2
+1	1	HD1 ND1 CG	CE1
+HSE		6		
+1	1	HN	N	-C	CA
+2	6	HB	CB	CG	CA
+1	5	HA	CA	N	C	CB
+1	1	HE2 NE2 CE1 CD2
+1	1	HD2 CD2 CG	NE2
+1	1	HE1 CE1 ND1 NE2
+HSP		7		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD2 CD2 CG	NE2
+1	1	HE1 CE1 ND1 NE2
+1	1	HD1 ND1 CG	CE1
+1	1	HE2 NE2 CE1 CD2
+HO4		1		
+3	10	HW	OW
+HO5		1		
+4	11	HW	OW
+HOH		1		
+2	7	HW	OW
+ILE		6		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+1	5	HB	CB	CA	CG1 CG2
+2	6	HG1 CG1 CD	CB
+3	4	HG2 CG2 CB	CA
+3	4	HD	CD	CG1 CB
+LEU		6		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	5	HG	CG	CB	CD1 CD2
+3	4	HD1 CD1 CG	CB
+3	4	HD2 CD2 CG	CB
+LSN		7
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	6	HD	CD	CE	CG
+2	6	HE	CE	NZ	CD
+2	4	HZ	NZ	CE	CD
+LYS		7		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	6	HD	CD	CE	CG
+2	6	HE	CE	NZ	CD
+3	4	HZ	NZ	CE	CD
+MET		5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	SD	CB
+3	4	HE	CE	SD	CG
+NME		2
+1	1	HN	N	-C	CH3
+3	4	HH3 CH3 N	-C
+PHE		8		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD1 CD1 CG	CE1
+1	1	HD2 CD2 CG	CE2
+1	1	HE1 CE1 CD1 CZ
+1	1	HE2 CE2 CD2 CZ
+1	1	HZ	CZ	CE1 CE2
+PRO		4		
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+2	6	HG	CG	CD	CB
+2	6	HD	CD	N	CG
+RA	9
+2	6	H5' C5' O5' C4'
+1	5	H4' C4' C5' O4' C3'
+1	5	H1' C1' O4' N9	C2'
+1	1	H8	C8	N9	N7
+2	3	H6	N6	C6	C5
+1	1	H2	C2	N1	N3
+1	5	H3' C3' C4' C2' O3'
+1	5	H2'' C2' C1' C3' O2'
+1	2	H2' O2' C2' C1'
+RU	9
+2	6	H5' C5' O5' C4'
+1	5	H4' C4' C5' O4' C3'
+1	5	H1' C1' O4' N1	C2'
+1	1	H6	C6	N1	C5
+1	1	H5	C5	C6	C4
+1	1	H3	N3	C4	C2
+1	5	H3' C3' C4' C2' O3'
+1	5	H2'' C2' C1' C3' O2'
+1	2	H2' O2' C2' C1'
+RG	9
+2	6	H5' C5' O5' C4'
+1	5	H4' C4' C5' O4' C3'
+1	5	H1' C1' O4' N9	C2'
+1	1	H8	C8	N9	N7
+1	1	H1	N1	C6	C2
+2	3	H2	N2	C2	N1
+1	5	H3' C3' C4' C2' O3'
+1	5	H2'' C2' C1' C3' O2'
+1	2	H2' O2' C2' C1'
+RC	9
+2	6	H5' C5' O5' C4'
+1	5	H4' C4' C5' O4' C3'
+1	5	H1' C1' O4' N1	C2'
+1	1	H6	C6	N1	C5
+1	1	H5	C5	C6	C4
+2	3	H4	N4	C4	C5
+1	5	H3' C3' C4' C2' O3'
+1	5	H2'' C2' C1' C3' O2'
+1	2	H2' O2' C2' C1'
+SER		4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	OG	CA
+1	2	HG1 OG	CB	CA
+SP1		4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	OG	CA
+1	2	HT	OT	P	OG
+SP2		3		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	OG	CA
+THP1	5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+1	5	HB	CB	CA	OG1 CG2
+1	2	HT	OT	P	OG1
+3	4	HG2 CG2 CB	CA
+THP2	4		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+1	5	HB	CB	CA	OG1 CG2
+3	4	HG2 CG2 CB	CA
+THR		5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+1	5	HB	CB	CA	OG1 CG2
+1	2	HG1 OG1 CB	CA
+3	4	HG2 CG2 CB	CA
+TP1		8		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD1 CD1 CG	CE1
+1	1	HD2 CD2 CG	CE2
+1	1	HE1 CE1 CD1 CZ
+1	1	HE2 CE2 CD2 CZ
+1	2	H2	O2	P1	OH
+TP2		7		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD1 CD1 CG	CE1
+1	1	HD2 CD2 CG	CE2
+1	1	HE1 CE1 CD1 CZ
+1	1	HE2 CE2 CD2 CZ
+TRP		9		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD1 CD1 NE1 CG
+1	1	HE1 NE1 CD1 CE2
+1	1	HE3 CE3 CD2 CZ3
+1	1	HZ3 CZ3 CE3 CH2
+1	1	HH2 CH2 CZ3 CZ2
+1	1	HZ2 CZ2 CE2 CH2
+TYR		8		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+2	6	HB	CB	CG	CA
+1	1	HD1 CD1 CG	CE1
+1	1	HD2 CD2 CG	CE2
+1	1	HE1 CE1 CD1 CZ
+1	1	HE2 CE2 CD2 CZ
+1	2	HH	OH	CZ	CE1
+URA 10 
+1	6	H5'  C5'  O5'  C4'
+1	6	H5'' C5'  C4'  O5'
+1	5	H4'  C4'  C5'  O4'	C3'
+1	5	H1'  C1'  O4'  N1	C2'
+1	1	H6	 C6   N1   C5
+1	1	H5	 C5   C6   C4
+1	1	H3	 N3   C4   C2
+1	5	H3'  C3'  C4'  C2'	O3'
+1	6	H2'' C2'  C1'  C3'
+1	2	H2'  O2'  C2'  C1'
+VAL		5		
+1	1	HN	N	-C	CA
+1	5	HA	CA	N	C	CB
+1	5	HB	CB	CA	CG1 CG2
+3	4	HG1 CG1 CB	CA
+3	4	HG2 CG2 CB	CA
