+++ gromacs_py/gmx/template/charmm36-jul2017.ff/merged.r2b
@@ -1,0 +1,16 @@
+
+
+; rtp residue to rtp building block table
+;GMX   Force-field
+HISD	HSD
+HIS1	HSD
+HISE	HSE
+HISH	HSP
+LYSN	LSN
+ASPH	ASPP
+GLUH	GLUP
+RA		ADE
+RG		GUA
+RU		URA
+RC		CYT
+
