+++ gromacs_py/gmx/template/charmm36-jul2017.ff/merged.vsd
@@ -1,0 +1,222 @@
+
+
+[ CH3 ]
+  CT3	  CE1	  MCH3
+  CT3	  C		  MCH3
+  CT3	  CA	  MCH3
+  CT3	  CC	  MCH3
+  CT3	  CD	  MCH3
+  CT3	  CPB	  MCH3
+  CT3	  CPH1	  MCH3
+  CT3	  CS	  MCH3
+  CT3	  CT1	  MCH3
+  CT3	  CT2	  MCH3
+  CT3	  CT3	  MCH3
+  CT3	  NC2	  MCH3
+  CT3	  NH1	  MCH3
+  CT3	  NH2	  MCH3
+  CT3	  NH3	  MCH3
+  CT3	  OC	  MCH3
+  CT3	  OH1	  MCH3
+  CT3	  OS	  MCH3
+  CT3	  S		  MCH3
+  CT3	  SM	  MCH3
+  CTL3	  CL	  MCH3 
+  CTL3	  OSL	  MCH3
+  CTL3	  CTL1	  MCH3
+  CTL3	  CTL2	  MCH3
+  CTL3	  CTL3	  MCH3
+  CTL3	  OHL	  MCH3
+  CTL3	  CEL1	  MCH3
+  CTL5	  NTL	  MCH3
+
+[ NH3 ]
+  NH3	  CT1	  MNH3
+  NH3	  CT2	  MNH3
+  NH3	  CT3	  MNH3
+  NH3L	  CTL2	  MNH3
+
+[ NH2 ]
+   NN1 planar
+   NH2	planar
+   NP	planar
+   NC2	planar
+
+
+; Data for generating dummy aromatic rings.
+; Actually we dont need all these bonds and angles,
+; but by specifying them here it is easier to improve
+; the dummy generation code later.
+[ PHE ]
+  CG	CD1		0.140
+  CG	CD2		0.140
+  CD1	CE1		0.140
+  CD2	CE2		0.140
+  CE1	CZ		0.140
+  CE2	CZ		0.140
+  CD1	HD1		0.108
+  CD2	HD2		0.108
+  CE1	HE1		0.108
+  CE2	HE2		0.108
+  CZ	HZ		0.108
+  CG	CD1   CE1	120.0
+  CD1	CE1   CZ	120.0
+  CE1	CZ	  CE2	120.0
+  CZ	CE2   CD2	120.0
+  CE2	CD2   CG	120.0
+  CD2	CG	  CD1	120.0
+  CG	CD1   HD1	120.0
+  CG	CD2   HD2	120.0
+  HD1	CD1   CE1	120.0
+  CD1	CE1   HE1	120.0
+  HE1	CE1   CZ	120.0
+  CE1	CZ	  HZ	120.0
+  HZ	CZ	  CE2	120.0
+  CZ	CE2   HE2	120.0
+  HE2	CE2   CD2	120.0
+  HD2	CD2   CG	120.0
+
+[ TYR ]
+  CG	CD1		0.140
+  CG	CD2		0.140
+  CD1	CE1		0.140
+  CD2	CE2		0.140
+  CE1	CZ		0.140
+  CE2	CZ		0.140
+  CD1	HD1		0.108
+  CD2	HD2		0.108
+  CE1	HE1		0.108	
+  CE2	HE2		0.108	
+  CZ	OH		0.1364	
+  OH	HH		0.0945
+  CG	CD1   CE1	120.0
+  CD1	CE1   CZ	120.0
+  CE1	CZ	  CE2	120.0
+  CZ	CE2   CD2	120.0
+  CE2	CD2   CG	120.0
+  CD2	CG	  CD1	120.0
+  CG	CD1   HD1	120.0
+  CG	CD2   HD2	120.0
+  HD1	CD1   CE1	120.0
+  CD1	CE1   HE1	120.0
+  HE1	CE1   CZ	120.0
+  CE1	CZ	  OH	120.0
+  CZ	OH	  HH	113.0
+  OH	CZ	  CE2	120.0
+  CZ	CE2   HE2	120.0
+  HE2	CE2   CD2	120.0
+  HD2	CD2   CG	120.0
+
+[ TRP ] ; angles from MD sim, corrected to be perfectly planar
+  CB	CG		0.1495
+  CG	CD1			0.1352
+  CD1	NE1		0.1381
+  NE1	CE2		0.138
+  CE2	CD2		0.1419
+  CD2	CG		0.1459
+  CE2	CZ2		0.140
+  CZ2	CH2		0.140
+  CH2	CZ3		0.140
+  CZ3	CE3		0.140
+  CE3	CD2		0.1404
+  CD1	HD1		0.108
+  NE1	HE1		0.101
+  CE3	HE3		0.108
+  CZ2	HZ2		0.108
+  CZ3	HZ3		0.108
+  CH2	HH2		0.108
+  CB	CG	  CD1		125.6
+  CB	CG	  CD2		128.7
+  CG	CD1   NE1	109.7
+  CD1	NE1   CE2	111.7
+  NE1	CE2   CD2	104.3
+  CE2	CD2   CG	108.5
+  CD2	CG	  CD1	105.7
+  NE1	CE2   CZ2	135.7
+  CE2	CZ2   CH2	120.0
+  CZ2	CH2   CZ3	120.0
+  CH2	CZ3   CE3	120.0
+  CZ3	CE3   CD2	120.0
+  CE2	CD2   CE3	120.0
+  CD2	CE2   CZ2	120.0
+  CE3	CD2   CG	131.5
+  CG	CD1   HD1	124.6
+  HD1	CD1   NE1	125.7
+  CD1	NE1   HE1	122.7
+  HE1	NE1   CE2	125.6
+  CZ3	CE3   HE3	120.0
+  HE3	CE3   CD2	120.0
+  CE2	CZ2   HZ2	120.0
+  HZ2	CZ2   CH2	120.0
+  CH2	CZ3   HZ3	120.0
+  HZ3	CZ3   CE3	120.0
+  CZ2	CH2   HH2	120.0
+  HH2	CH2   CZ3	120.0	
+
+[ HISA ] ; angles from MD sim, corrected to be perfectly planar
+  CG	ND1		0.1381
+  ND1	CE1		0.1343
+  CE1	NE2		0.1335
+  NE2	CD2		0.1394
+  CD2	CG		0.137
+  ND1	HD1		0.101
+  CE1	HE1		0.108
+  CD2	HD2		0.108
+  CG	ND1   CE1	107.4
+  ND1	CE1   NE2	113.1
+  CE1	NE2   CD2	104.5
+  NE2	CD2   CG	109.3
+  CD2	CG	  ND1	105.7
+  CG	ND1   HD1	127.7
+  HD1	ND1   CE1	124.9
+  ND1	CE1   HE1	122.9
+  HE1	CE1   NE2	124.0
+  CG	CD2   HD2	129.0
+  HD2	CD2   NE2	121.7
+
+[ HISB ] ; angles from MD sim, corrected to be perfectly planar
+  CG	ND1		0.1394
+  ND1	CE1		0.1335
+  CE1	NE2		0.1343
+  NE2	CD2		0.1381
+  CD2	CG		0.137
+  CD2	HD2		0.108
+  CE1	HE1		0.108
+  NE2	HE2		0.101
+  CG	ND1   CE1	104.9
+  ND1	CE1   NE2	112.9
+  CE1	NE2   CD2	107.3
+  NE2	CD2   CG	106.4
+  CD2	CG	  ND1	108.5
+  NE2	CD2   HD2	122.4
+  HD2	CD2   CG	131.2
+  ND1	CE1   HE1	123.8
+  HE1	CE1   NE2	123.3
+  CE1	NE2   HE2	125.9
+  HE2	NE2   CD2	126.8
+  
+[ HISH ] ; angles from MD sim, corrected to be perfectly planar
+  CG	ND1		0.1381
+  ND1	CE1		0.1343
+  CE1	NE2		0.1343
+  NE2	CD2		0.1381
+  CD2	CG		0.137
+  ND1	HD1		0.101
+  CE1	HE1		0.108
+  CD2	HD2		0.108
+  NE2	HE2		0.101
+  CG	ND1   CE1	107.4
+  ND1	CE1   NE2	111.4
+  CE1	NE2   CD2	107.2
+  NE2	CD2   CG	107.4
+  CD2	CG	  ND1	106.6
+  CG	ND1   HD1	127.4
+  HD1	ND1   CE1	125.2
+  ND1	CE1   HE1	124.3
+  HE1	CE1   NE2	124.3
+  CE1	NE2   HE2	126.4
+  HE2	NE2   CD2	126.4
+  NE2	CD2   HD2	120.8
+  HD2	CD2   CG	131.8
+
+
