+++ gromacs_py/gmx/template/charmm36-jul2017.ff/nbfix.itp
@@ -1,0 +1,31 @@
+; NBFIX  rmin=<charmm_rmin>/2^(1/6), eps=4.184*<charmm_eps>
+;name	type1  type2  1  sigma	 epsilon
+[ nonbond_params ]
+#ifndef USE_OLD_C36
+  NC2    OC  1  0.324019863788  0.648181096000
+#endif
+  SOD   CLA  1  0.332394311738  0.350933000000
+  POT   CLA  1  0.363575766873  0.477963424000
+  SOD   OCL  1  0.287760285959  0.313883680000
+  SOD   OBL  1  0.278851298778  0.313883680000
+  SOD   O2L  1  0.281523994932  0.313883680000
+  SOD    OC  1  0.287760285959  0.313883680000
+  SOD OC2D2  1  0.287760285959  0.313883680000
+  SOD OG2D2  1  0.287760285959  0.313883680000
+OG2D1 CLGR1  1  0.302905564168  0.836800000000
+    O CLGR1  1  0.302905564168  0.836800000000
+  ON1 CLGR1  1  0.302905564168  0.836800000000
+ ON1C CLGR1  1  0.302905564168  0.836800000000
+OC2D1 CLGR1  1  0.302905564168  0.836800000000
+ CRL1  HAL2  1  0.298451070577  0.134306400000
+ HGA1  HAL2  1  0.238760856462  0.150205600000
+ CRL2  HAL2  1  0.299341969295  0.173636000000
+ HGA2  HAL2  1  0.238760856462  0.132632800000
+ CRL1  HAL1  1  0.298451070577  0.122591200000
+ HGA1  HAL1  1  0.238760856462  0.137653600000
+ CRL2  HAL1  1  0.299341969295  0.158573600000
+ HGA2  HAL1  1  0.238760856462  0.120917600000
+ CRL1  CTL1  1  0.357250385974  0.115060000000
+ HGA1  CTL1  1  0.297560171859  0.128867200000
+ CRL2  CTL1  1  0.358141284692  0.148532000000
+ HGA2  CTL1  1  0.297560171859  0.113386400000
