+++ gromacs_py/gmx/template/charmm36-jul2017.ff/spc.itp
@@ -1,0 +1,30 @@
+
+[ moleculetype ]
+; molname	nrexcl
+SOL		2
+
+[ atoms ]
+;	nr	 type  resnr residue  atom	 cgnr	  charge	   mass
+	 1	   OW	1	 SOL	 OW		 1		-0.82
+	 2	   HW	1	 SOL	HW1		 1		 0.41
+	 3	   HW	1	 SOL	HW2		 1		 0.41
+
+#ifndef FLEXIBLE
+[ settles ]
+; OW	funct	doh dhh
+1	1	0.1 0.16330
+
+[ exclusions ]
+1	2	3
+2	1	3
+3	1	2
+#else
+[ bonds ]
+; i j	funct	length	force.c.
+1	2	1	0.1 345000	0.1		345000
+1	3	1	0.1 345000	0.1		345000
+	
+[ angles ]
+; i j	k	funct	angle	force.c.
+2	1	3	1	109.47	383 109.47	383
+#endif
